Add files via upload
Our code carries out two main functions: one is identifying contours, and the other is applying masks.
Masking- Masking is a process used in graphics software such as Photoshop to hide and reveal portions of an image. It is a non-destructive process of image editing.  One common method is to use a masking layer, which is essentially a second image that is used to cover up parts of the first image.
Contours - Contours can be explained simply as a curve joining all the continuous points along the boundary and having the same color or intensity. The contours are a useful tool for shape analysis, object detection, and recognition. Finding contours in OpenCV is comparable to finding a white object on a black background. Consequently, the background should be black and the object to be found should be white.